Correct Logo change
