+++ apps/console/app/components/ui/AgentLogo.tsx
@@ -1,13 +1,27 @@
-export function AgentLogo({ size = 32, ...props }: { size?: number }) {
+import type { ImgHTMLAttributes } from "react";
+
+type AgentLogoProps = Omit<
+  ImgHTMLAttributes<HTMLImageElement>,
+  "src" | "width" | "height"
+> & {
+  size?: number;
+  alt?: string;
+};
+
+export function AgentLogo({
+  size = 32,
+  alt = "Agent Logo",
+  ...props
+}: AgentLogoProps) {
   return (
     <img
       src="/hebo-icon.png"
-      alt="Agent Logo"
+      alt={alt}
+      {...props}
       width={size}
       height={size}
       loading="lazy"
       decoding="async"
-      {...props}
     />
   );
 }
